+++ packages/better-auth/src/api/rate-limiter/index.ts
@@ -1,6 +1,6 @@
-import type { AuthContext, RateLimit } from "../types";
-import { getIp } from "../utils/get-request-ip";
-import { logger } from "../utils/logger";
+import type { AuthContext, RateLimit } from "../../types";
+import { getIp } from "../../utils/get-request-ip";
+import { logger } from "../../utils/logger";
 
 function shouldRateLimit(
 	max: number,
