+++ packages/better-auth/src/init.ts
@@ -58,6 +58,11 @@ export const init = async (options: BetterAuthOptions) => {
 		baseURL: baseURL ? new URL(baseURL).origin : "",
 		basePath: options.basePath || "/api/auth",
 		plugins: plugins.concat(internalPlugins),
+		emailAndPassword: {
+			...options.emailAndPassword,
+			enabled: options.emailAndPassword?.enabled ?? false,
+			autoSignIn: true,
+		},
 	};
 	const cookies = getCookies(options);
